Move Frame alongside Scope.
Frame class is never used outside of scope, and is really just an
internal implementation detail...
